refactor: create LandingPageComponents
