+++ wealth-sync/src/app/page.tsx
@@ -1,10 +1,10 @@
 import Footer from "@/components/Footer/Footer";
-import IntroductionSection from "@/components/IntroductionSection";
-import FeaturesSection from "@/components/FeaturesSection";
+import IntroductionSection from "@/components/LandingPageComponents/IntroductionSection";
+import FeaturesSection from "@/components/LandingPageComponents/FeaturesSection";
 import IntegrationSection from "@/components/IntegrationPlatformsSection/IntegrationSection";
 // import PricingSection from "@/components/PricingPageComponents/PricingSection";
-import ChartPrievewSection from "@/components/ChartPrievewSection";
-import ActionSection from "@/components/ActionSection";
+import ChartPrievewSection from "@/components/LandingPageComponents/ChartPrievewSection";
+import ActionSection from "@/components/LandingPageComponents/ActionSection";
 import { NoPricingSection } from "@/components/PricingPageComponents/NoPricingSection/NoPricingSection";
 
 export default function HomePage() {
