+++ wealth-sync/src/components/LandingPageComponents/ActionSection.tsx
