+++ wealth-sync/src/components/LandingPageComponents/ChartPrievewSection.tsx
