+++ wealth-sync/src/components/LandingPageComponents/FeaturesSection.tsx
